Update applications.md
Slightly edit the application process.

NOTE: Omar suggested not limit the number of pages in the curriculum. He expects most students to have two or three pages at most, but if some student had some additional experience they may use the space.

I'll add the link to the PDF of the application form once we are happy with it.